Added http variables support to fuzz
